ci: update Dependabot ignores
Prevent Dependabot from updating Storybook to Storybook 9 as well as the
2 third party addons we use.

We cannot update yet because `@whitespace/storybook-addon-html` is not
yet compatible with Storybook 9.

Add `@whitespace/storybook-addon-html` to the npm-check-updates major
config as well.

Add comments as reminders to keep the dependabot ignores configuration
in line with the npm-check-updates configurations.